relocate core content to wallet core
