+++ packages/sdk/src/actions/swap/namespaces/WalletSwapNamespace.ts
@@ -1,7 +1,6 @@
 import { QUOTE_DISCRIMINATOR } from '@/actions/swap/core/SwapProvider.js'
 import { BaseSwapNamespace } from '@/actions/swap/namespaces/BaseSwapNamespace.js'
 import type { SupportedChainId } from '@/constants/supportedChains.js'
-import { executeTransactionBatch } from '@/core/executeTransactionBatch.js'
 import type { SwapExecuteParamsResolved } from '@/ens/types.js'
 import type { SwapSettings } from '@/types/actions.js'
 import type {
@@ -14,6 +13,7 @@ import type {
 } from '@/types/swap/index.js'
 import type { TransactionData } from '@/types/transaction.js'
 import type { RecipientResolver } from '@/utils/ens.js'
+import { executeTransactionBatch } from '@/wallet/core/utils/executeTransactionBatch.js'
 import type { Wallet } from '@/wallet/core/wallets/abstract/Wallet.js'
 
 /**
