+++ packages/sdk/src/wallet/core/error/errors.ts
