+++ packages/sdk/src/wallet/core/wallets/smart/default/DefaultSmartWallet.ts
@@ -11,13 +11,13 @@ import type { WaitForUserOperationReceiptReturnType } from 'viem/account-abstrac
 import { toCoinbaseSmartAccount } from 'viem/account-abstraction'
 
 import type { SupportedChainId } from '@/constants/supportedChains.js'
-import { TransactionConfirmedButRevertedError } from '@/core/error/errors.js'
-import { retryOnStaleRead } from '@/core/utils/retryOnStaleRead.js'
 import type { ChainManager } from '@/services/ChainManager.js'
 import type { Asset } from '@/types/asset.js'
 import type { LendProviders, SwapProviders } from '@/types/providers.js'
 import type { TransactionData } from '@/types/transaction.js'
 import { parseAssetAmount } from '@/utils/assets.js'
+import { TransactionConfirmedButRevertedError } from '@/wallet/core/error/errors.js'
+import { retryOnStaleRead } from '@/wallet/core/utils/retryOnStaleRead.js'
 import { SmartWallet } from '@/wallet/core/wallets/smart/abstract/SmartWallet.js'
 import type { Signer } from '@/wallet/core/wallets/smart/abstract/types/index.js'
 import {
